Problems

Problems

A Problem is the structure representing the problem that must be solved, such as the minimization of the cost function to determine the steady state. Currently only problems representing the computation of the steady state by minimization of the cost function $\mathcal{L}^\dagger\mathcal{L}$ are supported.

Minimization of $\mathcal{C} = \langle\langle\mathcal{L}^\dagger\mathcal{L}\rangle\rangle$

This cost function can be computed in two ways:

\[\mathcal{C} = \sum_\sigma p(\sigma) \langle\langle\sigma |\mathcal{L}^\dagger\mathcal{L}\rho\rangle\rangle\]

or

\[\mathcal{C} = \sum_\sigma p(\sigma) |\langle\langle\sigma |\mathcal{L}\rho\rangle\rangle||^2\]

The second version leads to smaller variance of sampled variables and also is faster to evaluate because it holds only $\mathcal{L}$ instead of $\mathcal{L}^\dagger\mathcal{L}$, as such I reccomend to use this one.

NeuralQuantum.LdagL_Lmat_prob — Type.

LdagL_Lmat_prob <: Problem

Problem or finding the steady state of a ℒdagℒ matrix by computing 𝒞 = ∑|ρ(σ)|²|⟨⟨σ|ℒ |ρ⟩⟩|² only storing H and c_ops.

DO NOT USE WITH COMPLEX-WEIGHT NETWORKS, AS IT DOES NOT WORK

source

NeuralQuantum.LdagL_spmat_prob — Type.

LdagL_spmat_prob <: Problem

Problem or finding the steady state of a ℒdagℒ matrix

source