Problems

A Problem encodes a many-body quantum problem into a cost function to be minimized. Problems can encode the search for the Ground State energy of an hamiltonian of for the Steady-State of an Open Quantum System.

Steady States of Open Quantum Systems

A SteadyStateProblem encodes the steady-state of a liouvillian into the global minimum of the following cost function

$\mathcal{C} = \langle\langle\mathcal{L}^\dagger\mathcal{L}\rangle\rangle$

When stochastically sampling this cost function on the hilbert space, we can evaluate it in two possible ways:

\[\mathcal{C} = \sum_\sigma p(\sigma) \langle\langle\sigma |\mathcal{L}^\dagger\mathcal{L}\rho\rangle\rangle\]

or (variance style)

\[\mathcal{C} = \sum_\sigma p(\sigma) |\langle\langle\sigma |\mathcal{L}\rho\rangle\rangle|^2\]

The second version leads to smaller variance of sampled variables and also is faster to evaluate because it holds only $\mathcal{L}$ instead of $\mathcal{L}^\dagger\mathcal{L}$, as such I reccomend to use this one. To use it, set variance=true when costructing the problem.

NeuralQuantum.SteadyStateProblem — Function

SteadyStateProblem([T=STD_REAL_PREC], ℒ, operators=true, variance=true, superop=false)
SteadyStateProblem([T=STD_REAL_PREC], H, J, operators=true, variance=true)

Returns the problem targeting the steady-state of the Liouvillian ℒ through a minimisation of the cost function 𝒞 = Tr[ρ'ℒ'ℒ ρ]/Tr[ρ'ρ] .

If variance=true the cost function is sampled by evaluating ℒρ, which has better convergence properies. The sampling is performed on ℒ'ℒ ρ otherwise.

See appendix of https://arxiv.org/abs/1902.10104 for more info.

If superop=false a Klocal operator of the lindbladian is used (unstable).

If operators=true a memory efficient representation of the hamiltonian is used, resulting in less memory consuption but higher CPU usage. This is needed for lattices bigger than a certain threshold. To use this feature, the hamiltonian must be provided as a GraphOperator object, from QuantumLattices package.

Ground State Problems

A GroundStateProblem encodes the ground-state of an Hamiltonian into the global minimum of the total energy

$\mathcal{C} = \langle\psi|\hat{H}|\psi\rangle$

To construct a GroundStateProblem you must supply it a valid hamiltonian.

NeuralQuantum.GroundStateProblem — Function

GroundStateProblem([T=STD_REAL_PREC], ham, operators=true, variance=false)

Returns the problem targeting of the ground state energy. If variance=false this is done by minimizing the energy, otherwise it's done by minimizing the variance.

If operators=true a memory efficient representation of the hamiltonian is used, resulting in less memory consuption but higher CPU usage. This is needed for lattices bigger than a certain threshold. To use this feature, the hamiltonian must be provided as a GraphOperator object, from QuantumLattices package.