Reference

This is the reference to the public interface of NeuralQuantum.

Hilbert spaces

Methods relative to Hilbert spaces and their states

General methods

NeuralQuantum.nsites — Function

nsites(hilbert) -> Int

Returns the number of lattice sites in the Hilbert space.

NeuralQuantum.shape — Function

shape(hilbert) -> Vector{Int}

Returns a vector containing the local hilbert space dimensions of every mode in the Hilbert space.

In the case of homogeneous spaces, it is usually more efficient to call local_dim.

NeuralQuantum.local_dim — Function

local_dim(hilbert [, i]) -> Int

Returns the local dimension of the hilbert space on site i. If the space is homogeneous then specifying i is not needed.

NeuralQuantum.spacedimension — Function

spacedimension(hilbert) = prod(shape(hilbert)) -> Int

Returns the total dimension of the vector space hilbert. This is only valid if indexable(hilbert) == true, otherwise it's 0.

NeuralQuantum.indexable — Function

indexable(hilbert) -> Bool

Returns true if the space is can be indexed with an Int64, which means that htis tests true when

    spacedimension(hilbert) <= typemax(Int64)

NeuralQuantum.is_homogeneous — Function

is_homogeneous(hilbert) -> Bool

Returns true if the space is homogeneous, that is, if all the modes have the same local hilbert space dimension.

Constructors

NeuralQuantum.HomogeneousFock — Type

HomogeneousFock(N, m)

Constructs the Hilbert space of N identical modes with m levels.

NeuralQuantum.HomogeneousSpin — Type

HomogeneousSpins(N, S::Rational=1//2)

Constructs the Hilbert space of N identical spins-S (by default 1//2).

Working with states

NeuralQuantum.index — Function

index(hilb, state)

Converts to an int the state of the hilbert space hilb assuming it's indexable.

NeuralQuantum.state — Function

state([arrT=Vector], [T=STD_REAL_PREC], hilbert)

Constructs a state for the hilbert space with precision T and array type arrT.

NeuralQuantum.states — Function

states(hilb) -> iterator

Returns an iterator to iterate all states in the hilbert space

Missing docstring.

Missing docstring for Random.rand!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Random.rand. Check Documenter's build log for details.

NeuralQuantum.apply! — Function

apply!(σ, changes)

Applies the changes changes to the σ.

If state isa DoubleState then single-value changes are applied to the columns of the state (in order to compute matrix-operator products). Otherwise it should be a tuple with changes of row and columns.

If changes is nothing, does nothing.

apply!(state, changes, [changes_r] )

Apply the changes changes to the state. If state is a double state and changes is a tuple, then the first element of the tuple is the row-changes and the second element is the column-changes. Optionally the two elements of the tuple can be passed separately.

If the state is double but there is only 1 element of changes, it's applied to the rows.

NeuralQuantum.flipat! — Function

flipat!(state, hilb, i) -> (old_val, new_val)

Randomly flips state[i] to another available state. Returns the old value of state[i] and the new value. state is changed in-place.

Missing docstring.

Missing docstring for NeuralQuantum.setat!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuralQuantum.local_index. Check Documenter's build log for details.

NeuralQuantum.unsafe_get_el — Function

unsafe_get_el(state, [batch], el)

Given a vector of batches of states with size size(state) = [:, batches, els], take the batch group el, and if specified also selects one single batch.

It's somewhat equivalent to a view, but handles tuples of states for density matrices correctly and uses unsafe views to prevent allocation on CPU.

Operators

QuantumOpticsBase.sigmax — Function

sigmax(hilbert::AbstractHilbert, site::Int)

Builds a σₓ operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: M-dimensional Hilbert spaces are treated as spin (M-1)//2 spins.

QuantumOpticsBase.sigmay — Function

sigmay(hilbert::AbstractHilbert, site::Int)

Builds a σ_y operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: Hilbert must have local space dimension 2 on that site.

QuantumOpticsBase.sigmaz — Function

sigmaz(hilbert::AbstractHilbert, site::Int)

Builds a σ_z operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: M-dimensional Hilbert spaces are treated as spin (M-1)//2 spins.

QuantumOpticsBase.sigmap — Function

sigmap(hilbert::AbstractHilbert, site::Int)

Builds a σ₊ operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: M-dimensional Hilbert spaces are treated as spin (M-1)//2 spins.

QuantumOpticsBase.sigmam — Function

sigmam(hilbert::AbstractHilbert, site::Int)

Builds a σ₋ operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: M-dimensional Hilbert spaces are treated as spin (M-1)//2 spins.

QuantumOpticsBase.create — Function

create(hilbert::AbstractHilbert, site::Int)

Builds a bosonic creation operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: spin-m//2 spaces are treated as fock spaces with dimension m+1

QuantumOpticsBase.destroy — Function

destroy(hilbert::AbstractHilbert, site::Int)

Builds a bosonic destruction operator acting on the site-th site of the Many body Hilbert space hilbert.

Note: spin-m//2 spaces are treated as fock spaces with dimension m+1

NeuralQuantum.KLocalOperator — Type

KLocalOperator

A representation of a local operator, acting on certain sites, each with hilb_dims dimension. The local operator is written in matrix form in this basis as mat. For every row of mat there are several non-zero values contained in mel, and to each of those, to_change contains the sites that must change the basis new value, contained in new_value

NeuralQuantum.KLocalOperatorRow — Function

KLocalOperatorRow(sites::Vector, hilb_dims::Vector, operator)

Creates a KLocalOperator where connections are stored by row for the operator operator acting on sites each with hilb_dims local dimension.

QuantumOpticsBase.liouvillian — Function

liouvillian(Ĥ, ops::Vector)

Constructs the LocalOperator representation of a liouvillian super-operator with coherent part given by the hamilotnian Ĥ, and ops as the list of jump operators.

Samplers

NeuralQuantum.ExactSampler — Type

ExactSampler(n_samples; seed=rand)

Constructs an exact sampler, which builds the full pdf of the quantum state and samples it exactly.

This sampler can only be used on indexable spaces, and should be used only for somewhat small systems (Closed N<14, Open N<7), as the computational cost increases exponentially with the number of sites.

Initial seed can be set bu specifying seed.

NeuralQuantum.MetropolisSampler — Type

MetropolisSampler(rule, chain_length, passes; burn=0, seed=rand)

Constructs a Metropolis-Hastings sampler which samples Markov chains of length chain_length + burn, ignoring the first burn samples. Transition rules are specified by the rule rule. To reduce auto-correlation, passes number of metropolis-hastings steps are performed for each sample returned (minimum 1).

Effectively, this means that the chain is actually passes * (chain_lenght + burn) long, but only 1 every passes elements are stored and used to compute expectation values.

Initial seed can be set bu specifying seed.

Transition Rules

NeuralQuantum.LocalRule — Type

LocalRule()

Transition rule for Metropolis-Hastings sampling where at every step a random site is switched to another random state.

NeuralQuantum.ExchangeRule — Type

ExchangeRule(graph)

Transition rule for Metropolis-Hastings sampling where at every step a random couple of sites i,j is switched.

Couples of sites are generated from the graph graph (or operator), where all sites that are connected (or coupled by a 2-body term) are considered for switches.